Use omp_get_wtime to calculate the execution time of the sequential program

+++ pthread/kmeans-sequential/include/helper.h
@@ -14,7 +14,7 @@ typedef struct {
 
 typedef struct {
     int number;
-    float executionTime;
+    double executionTime;
     int convergenceIteration;
 } Experiment;
